Update Pictures
I added new pictures and changed it so when you click on a picture, it
would show the image in full, rather than the default image. I also
commented out the titles because I thought they looked better that way,
but left them as comments so we can add them back in if we want to use
them in the future. I just thought the titles got in the way of viewing
the photo too much.